Adds support for power-of-2 decimation for both real and complex sample buffers.  Creates new class to analyze the filter attenuation rates using a variety of filter lengths and window options to enable selecting the optimal combination of cascaded power-of-2 half band decimation filters. (#1127)

+++ src/main/java/io/github/dsheirer/dsp/filter/FilterFactory.java
@@ -1004,7 +1004,7 @@ public class FilterFactory
      * @param windowType to apply to the filter.
      * @return coefficients
      */
-    public static double[] getHalfBand(int length, Window.WindowType windowType)
+    public static float[] getHalfBand(int length, Window.WindowType windowType)
     {
         if((length - 3) % 4 != 0)
         {
@@ -1013,7 +1013,7 @@ public class FilterFactory
         }
 
         double[] window = Window.getWindow(windowType, length);
-        double[] taps = new double[length];
+        float[] taps = new float[length];
 
         int halfLength = length / 2;
 
@@ -1053,7 +1053,7 @@ public class FilterFactory
 
         int length = 15;
 
-        double [] taps = FilterFactory.getHalfBand(length, Window.WindowType.HAMMING);
+        float [] taps = FilterFactory.getHalfBand(length, Window.WindowType.HAMMING);
 
         for(int x = 0; x < length; x++)
         {
